CDB_HexagonGrid: take tiling into account
This commit adds an optional "origin" parameter to the function
which defaults to (0,0) allowing control over the grid origin and
thus letting you produce big coverages using multiple small calls.

NOTE: snapping to a precision grid for perfect side matching is
still to be implemented. Why ? Take a look at this:
http://trac.osgeo.org/postgis/ticket/1789