Set ST_MakeValid call timeout to 10mins (#730)
There's another 10mins timeout on the CDB_TransformToWebmercator.
In both cases the slow part is ST_MakeValid, used by the latter
but only for those geometries which fall outside of the valid
webmercator bounds. Theoretically the CDB_TransformToWebmercator
timeout could be lower, but isn't worth messing with timeouts
which shouldn't be there in the first place.