Some fixes in user relocation scripts
Added methods to return the db name and username prefixes depending on
the running environment.
Don't rename db username in origin when performing the db dump.
Replace db name and username in sql dump by the ids of the destination
user.
Convert privacy value to integer before inserting it in the db.
Fixed the method that returns from redis all the tables in a database.