1037 Modify QueryBatcher to not need an id_column
Modify the QueryBatcher so that it does not require an id_column
parameter. Two different ID's were used, depending on the status of the
table: cartodb_id and ogc_fid.

Now it uses the CTID, that is present on all tables.

See http://www.postgresql.org/docs/9.3/static/ddl-system-columns.html
for more info.