Only add layer once confirmed it was added
Previously even if the layer failed to be saved for whatever reason it
was still added to the collection and thus appeared in the map’s
collection. Using create and the wait: true prop it won’t add it until
the server-side responded correctly.

The server-side issue of layer being saved in inconsistent state still
needs to be fixed in addition to this.