Implement (single) date picker
Based on the already existing dates range picker. It was was renamed
and moved to a common place, since they are both related, and since now
having a “single date picker” replaced its conceptual name. Also had to
update styles, to not mess with existing styles that was done as
default for the date range picker I added new rules that overrides the
defaults instead