change default Torque color to #0F3B82
Given that we have the light basemap, the default torque color isn't cutting it anymore. I'm changing it here to a dark blue. #0F3B82. Combined with the blend mode (lighter) it also gives a really nice intensity effect but still pops on the map when there are few circles.

cc @saleiva @javierarce (this should likely be the same setting for pecan)