Move expanding rule to an inner element
This is due to a block (flex or not) with an overflow do not apply the
paddings, as explained in the stack overflow link in the inline
comment. The fix is to move the rule to an inner element where the
padding actually is applied, so it works consistently for FF and other
browsers.