Add group users directly from empty group
Solved by having a parent view “on top” which manages the edge-cases of
transitioning from/to empty group (i.e. no group users).

Also had to modify the group users collection to do a fetch on the
batch operations to reflect the state (thus the promise resolves are
slightly delayed)