Improve empty CSV files handling
For empty CSVs, which usually came from unsupported Excel files, such
as sheets with graphs, the normalization was crashing.

The first_line method was returning the `stream.rewind` value in case
that the file was empty, and therefore the first_line would be nil.

This commit also edits the CSV column number checking in order to
be able to admit empty files (otherwise, nil.length would raise an
exception).