Clean final breaklines from SQL geocoder output
Fixes #6532. For single-line CSVs, the parser does not handle correctly
the final \r\n that the SQL API returns. This code chomps the response
so that it can get rid of this \r\n.

This also includes a minor improvement, checking `response.blank?`
instead of `response = '\n'` to check if it's empty or not.