Avoid table names prefixed by an underscore and reserve public
Through the different pieces of the platform, underscored table names
have been causing issues. (i.e. sequel is crashing when georeferencing a
table whose name starts by an underscore). Although Postgres supports
these names, for each table it creates an internal type preceded by an
underscore. Therefore, there are collisions between underscored names
and already existent types.

This commit will rename any table whose name starts by an underscore.
Besides we have added the word "public" to the array of reserved words,
because it is not officially a reserved word in postgres but it can
cause issues and it is a reserved word in other standards.