Disable PROMOTE TO MULTI option when CSV guessing
Previously, all the output from ogr2ogr was promoted to multi. Part of
this is handled in georeferencer.rb and column.rb, but the scenario for
CSV guessing from which a point was created from lat/long columns was
bringing MultiPoints to CartoDB tables.

Ogr2ogr was returning a wkb_column(geometry(MultiPoint, 4326)) that the
CartoDBfication process, at this time, would always use, discarding
another existent the_geom column in the table. This wkb_geometry was not
treated to make it Point in Rails because the code here is taking care
of the_geom columns, so the issue happens in combination with the
CartoDBfication geometry column selection.