Add notify debug when gdrive item structure not as expected
Ignores invalid file in gdrive list. In issue 6854 we noticed that there
are items retrieved from Google Drive that do not match with the two
kind of files we are expecting and make the list of gdrive to crash,
making impossible to the affected users to use the gdrive importer. This
commits adds a debugger functionality to these cases and ignores them to
workaround the issue while we found the final solution.