Basic implementation of odbc/fdw connectors
This adds supports for using odbc fdw in imports and synchronizations
to import datasets directly from external database queries.

This requires, in the db server host, the installation of odbc_fdw
and odbc drivers for the sources to be used.
The CartoDB fork of odbc_fdw  must be used, which fixes some bugs,
adds support for PostgreSQL 9.5 and allows passing DRIVER,
HOST and PORT parameters without requiring a DSN.

For this proof-of-concept implementation, an odbc parameters
is passed to the import API, encoding the necessary parameters
(driver, host, port, username, password, database, table, sql_query,
sql_count and columns) into an string of the form "par=val;...".

The characters '=' and ';' can be encoded as '.eq.' and '.sc.' in
sql_query and sql_count.

The columns parameter is currently necessary to define the columns
of the table that will be imported in PostgreSQL syntax.

For DataImport and Synchronization records created using the odbc parameter,
the value 'connector' is stored in column service_name and a
connection string of the form. "ODBC:par=val;..." is stored in service_item_id.
This is done provisionally to avoid modifying the db schema while testing
this feature. Proper columns should be added to DataImport and Synchronization.