Downsample raster overviews
Downsample original raster before proceeding with the rest of the
process of overview generation. Unfortunately this cannot be done along
with the reprojection as gdalwarp command does not support the `-scale`
parameter of gdal_translate, and conversely gdal_translate does not
support reprojection.

This can be done better but for the moment this should fit the bill.