Require a feature-flag "carto-connectors" to activate connectors
Currently the feature flag is checked during the execution of
an import/synchronization. It might be desirable to have the
creation of the Import/Synchronization fail when connectors are
not available.

Also note that when an Import fails that has a related Sychronization,
a Synchronization which will never be executed is left in the DB
(as occurs with other types of failed Imports).