Implement a buffer to avoid race conditions.
I've observed styles changes a lot when the user changes the aggregation carousel. Because a model
is not added to the collection until it's saved in the backend, we save the model in a buffer
meanwhile. So we checked if a model of such type for that layer exists in the legend collection or
in the buffer before to make the request.