Make sure pre_client task takes care of dashboard files.
Not sure if this task should run as part of the dev task. It asumes all files are in place, but the
thing is if you come from an older topic branch mergin master, this files are not going to be there.
And you must to run pre_client task before dev.