Rebind events.
Notifier view is a singleton view. This means, the view is created once, then added and removed.
When removed, because the internal management of our views, we destroy the bindings. This happens
when we change to the options view. The notifier parent's view, editor-map-view, is deleted, so ir
permorfs a clearSubViews, deleting the bindings. In order to fix this, every time we append the
notifier view, we need to make sure bindings are working properly.