Allow to override S3 endpoint
For visualization exports and data imports it's now possible to override
the S3 endpoint which by default points to AWS URLs. The config
parameter is called s3_endpoint. This is useful when a user, in a
on-premise installation, is working with a S3 compatible storage server
different than AWS one