Get rid of get_new_system_geocoding_calls
In the end very similar or even exactly the same implementation of
get_geocoding_calls and not used in production. The only 2 usages were
from rake tasks. One of them was to check the "old" against the "new"
function and the other one to produce a report that has not been
exercised in a long while.